requirements.txt
pandas>=2.1
numpy>=1.24
pyarrow>=14.0
PyYAML>=6.0
pytest>=7.4
